Split out lib functions into their own file
